refactored redux module into redux-config
