+++ modules/redux-config/index.js
@@ -1,0 +1,6 @@
+import reducers from './reducers'
+import initialState from './initialState'
+import initStore from './store'
+
+// expose a function for pages to grab/init the store
+export const getStore = () => initStore(reducers, initialState)
