+++ modules/redux-config/initialState.js
@@ -1,0 +1,8 @@
+// TODO - this info should really reside in their own respective
+// feature modules, refactor this in the future
+
+// this initial state is used for server-side rendering as well
+export default {
+  counter: { count: 3 },
+  todoapp: { count: -3 },
+}
