+++ modules/redux-config/reducers.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux'
+import { reducer as counter } from '~/counter'
+import { reducer as todoapp } from '~/todoapp'
+
+export default combineReducers({
+  counter,
+  todoapp
+})
