+++ modules/redux-config/store.js
@@ -1,0 +1,12 @@
+import { createStore } from 'redux'
+import { composeWithDevTools } from 'redux-devtools-extension';
+
+let store = null
+
+export default (reducer, initialState = {}, isServer) => {
+  const enhancers = composeWithDevTools()
+  if (!store) {
+    store = createStore(reducer, initialState, enhancers)
+  }
+  return store ? store : createStore(reducer, initialState, enhancers)
+}
